refactor(agent): simplify addressed session lifecycle
Reuse canonical identity helpers, remove redundant pane caching, narrow hydration to the active inventoried pane, stabilize refresh without coupling pane props to the session API object, and decompose EmbeddedAgentGateway below 1k lines.
